refactor: replace hash-wasm with custom hashing functions and remove unused files
